rework how we get hold of intercepted actors
