+++ Sources/DistributedActors/ClusterSystem.swift
@@ -863,12 +863,12 @@ extension ClusterSystem {
         where Act: DistributedActor
     {
         self.log.trace("Resolve: \(id)")
-        
+
         if let interceptor = id.context.remoteCallInterceptor {
             self.log.trace("Resolved \(id) as intercepted", metadata: ["interceptor": "\(interceptor)"])
             return nil
         }
-        
+
         guard self.cluster.uniqueNode == id.uniqueNode else {
             self.log.trace("Resolved \(id) as remote, on node: \(id.uniqueNode)")
             return nil
@@ -902,16 +902,16 @@ extension ClusterSystem {
     {
         return self._assignID(actorType, baseContext: nil)
     }
-    
+
     public func _assignID<Act>(_ actorType: Act.Type, baseContext: DistributedActorContext?) -> ClusterSystem.ActorID
         where Act: DistributedActor
     {
         let props = _Props.forSpawn // task-local read for any properties this actor should have
-        
+
         if let designatedActorID = props._designatedActorID {
             return designatedActorID
         }
-        
+
         var id = try! self._reserveName(type: Act.self, props: props)
 
         let lifecycleContainer: LifecycleWatchContainer?
@@ -986,23 +986,23 @@ extension ClusterSystem {
 // MARK: Intercepting calls
 
 extension ClusterSystem {
-    
     public func interceptCalls<Act, Interceptor>(
         to actorType: Act.Type,
         metadata: ActorMetadata,
-        interceptor: Interceptor) throws -> Act
-    where Act: DistributedActor, Act.ActorSystem == ClusterSystem,
-          Interceptor: RemoteCallInterceptor
+        interceptor: Interceptor
+    ) throws -> Act
+        where Act: DistributedActor, Act.ActorSystem == ClusterSystem,
+        Interceptor: RemoteCallInterceptor
     {
         /// Prepare a distributed actor context base, such that the reserved ID will contain the interceptor in the context.
         let baseContext = DistributedActorContext(lifecycle: nil, remoteCallInterceptor: interceptor)
         var id = self._assignID(Act.self, baseContext: baseContext)
         assert(id.context.remoteCallInterceptor != nil)
         id = id._asRemote // FIXME(distributed): not strictly necessary ???
-        
+
         var props = _Props()
         props._designatedActorID = id
-        
+
         return try Act.resolve(id: id, using: self)
     }
 }
@@ -1237,22 +1237,20 @@ public struct ClusterInvocationResultHandler: DistributedTargetInvocationResultH
         )
         case localDirectReturn(CheckedContinuation<Any, Error>)
     }
-    
-    
+
     init(system: ClusterSystem, callID: ClusterSystem.CallID, channel: Channel, recipient: ClusterSystem.ActorID) {
         self.state = .remoteCall(system: system, callID: callID, channel: channel, recipient: recipient)
     }
-    
+
     init(directReturnContinuation: CheckedContinuation<Any, Error>) {
         self.state = .localDirectReturn(directReturnContinuation)
     }
-    
 
     public func onReturn<Success: Codable>(value: Success) async throws {
         switch self.state {
         case .localDirectReturn(let directReturnContinuation):
             directReturnContinuation.resume(returning: value)
-            
+
         case .remoteCall(_, let callID, let channel, let recipient):
             let reply = RemoteCallReply<Success>(callID: callID, value: value)
             try await channel.writeAndFlush(TransportEnvelope(envelope: Payload(payload: .message(reply)), recipient: recipient))
@@ -1263,7 +1261,7 @@ public struct ClusterInvocationResultHandler: DistributedTargetInvocationResultH
         switch self.state {
         case .localDirectReturn(let directReturnContinuation):
             directReturnContinuation.resume(returning: ())
-            
+
         case .remoteCall(_, let callID, let channel, let recipient):
             let reply = RemoteCallReply<_Done>(callID: callID, value: .done)
             try await channel.writeAndFlush(TransportEnvelope(envelope: Payload(payload: .message(reply)), recipient: recipient))
@@ -1274,7 +1272,7 @@ public struct ClusterInvocationResultHandler: DistributedTargetInvocationResultH
         switch self.state {
         case .localDirectReturn(let directReturnContinuation):
             directReturnContinuation.resume(throwing: error)
-            
+
         case .remoteCall(let system, let callID, let channel, let recipient):
             system.log.warning("Result handler, onThrow: \(error)")
             let reply: RemoteCallReply<_Done>
