+++ Sources/DistributedActors/Props.swift
@@ -54,7 +54,7 @@ public struct _Props: @unchecked Sendable {
     /// Used only with 'distributed actor' as a way to pass path to the `assignIdentity` call.
     /// // TODO(distributed): We should instead allow for an explicit way to pass params to the transport.
     internal var _knownActorName: String?
-    
+
     /// Makes `DistributedActor.resolve` use the designated ID, rather than assigning one.
     internal var _designatedActorID: ActorID?
 
@@ -183,7 +183,7 @@ extension _Props {
         props._knownActorName = name
         return props
     }
-    
+
     public func _knownAs(name: String) -> Self {
         var p = self
         p._knownActorName = name
