+++ Tests/DistributedActorsTests/InterceptorTests.swift
@@ -12,6 +12,7 @@
 //
 //===----------------------------------------------------------------------===//
 
+import Distributed
 @testable import DistributedActors
 import DistributedActorsTestKit
 import Foundation
@@ -56,6 +57,54 @@ final class TerminatedInterceptor<Message: Codable>: _Interceptor<Message> {
 }
 
 final class InterceptorTests: ClusterSystemXCTestCase {
+    func test_remoteCall_interceptor() async throws {
+        let local = await setUpNode("local") { settings in
+            settings.enabled = true
+        }
+        let remote = await setUpNode("remote") { settings in
+            settings.enabled = true
+        }
+        local.cluster.join(node: remote.cluster.uniqueNode)
+
+        let otherGreeter = Greeter(actorSystem: local, greeting: "HI!!!")
+        let localGreeter: Greeter = try system.interceptCalls(
+            to: Greeter.self,
+            metadata: ActorMetadata(),
+            interceptor: GreeterRemoteCallInterceptor(system: local, greeter: otherGreeter)
+        )
+
+        let value = try await shouldNotThrow {
+            try await localGreeter.greet()
+        }
+        value.shouldEqual("HI!!!")
+    }
+
+    func test_remoteCallVoid_interceptor() async throws {
+        let local = await setUpNode("local") { settings in
+            settings.enabled = true
+        }
+        let remote = await setUpNode("remote") { settings in
+            settings.enabled = true
+        }
+        local.cluster.join(node: remote.cluster.uniqueNode)
+
+        let otherGreeter = Greeter(actorSystem: local, greeting: "HI!!!")
+        let localGreeter: Greeter = try shouldNotThrow {
+            try system.interceptCalls(
+                to: Greeter.self,
+                metadata: ActorMetadata(),
+                interceptor: GreeterRemoteCallInterceptor(system: local, greeter: otherGreeter)
+            )
+        }
+
+        try await shouldNotThrow {
+            try await localGreeter.muted()
+        }
+        try self.capturedLogs(of: local).awaitLogContaining(self.testKit(local), text: "Muted greeting: HI!!!")
+    }
+
+    // Legacy interceptor API tests -----------------------------------------------------------------------------------
+
     func test_interceptor_shouldConvertMessages() throws {
         let p: ActorTestProbe<String> = self.testKit.makeTestProbe()
 
@@ -188,3 +237,125 @@ final class InterceptorTests: ClusterSystemXCTestCase {
         try p.expectTerminated(ref)
     }
 }
+
+private distributed actor Greeter {
+    typealias ID = ClusterSystem.ActorID
+    typealias ActorSystem = ClusterSystem
+
+    let greeting: String
+
+    init(actorSystem: ActorSystem, greeting: String) {
+        self.actorSystem = actorSystem
+        self.greeting = greeting
+    }
+
+    distributed func greet() async throws -> String {
+        self.greeting
+    }
+
+    distributed func greetThrow(codable: Bool) async throws -> String {
+        if codable {
+            throw GreeterCodableError()
+        } else {
+            throw GreeterNonCodableError()
+        }
+    }
+
+    distributed func greet(delayNanos: UInt64) async throws -> String {
+        try await Task.sleep(nanoseconds: delayNanos)
+        return try await self.greet()
+    }
+
+    distributed func muted() async throws {
+        self.actorSystem.log.info("Muted greeting: \(self.greeting)")
+    }
+
+    distributed func mutedThrow(codable: Bool) async throws {
+        if codable {
+            throw GreeterCodableError()
+        } else {
+            throw GreeterNonCodableError()
+        }
+    }
+
+    distributed func muted(delayNanos: UInt64) async throws {
+        try await Task.sleep(nanoseconds: delayNanos)
+        try await self.muted()
+    }
+}
+
+private struct GreeterRemoteCallInterceptor: RemoteCallInterceptor {
+    let system: ClusterSystem
+    let greeter: Greeter
+
+    init(system: ClusterSystem, greeter: Greeter) {
+        self.system = system
+        self.greeter = greeter
+    }
+
+    func interceptRemoteCall<Act, Err, Res>(
+        on actor: Act,
+        target: RemoteCallTarget,
+        invocation _invocation: inout ClusterSystem.InvocationEncoder,
+        throwing: Err.Type,
+        returning: Res.Type
+    ) async throws -> Res
+        where Act: DistributedActor,
+        Act.ID == ActorID,
+        Err: Error,
+        Res: Codable
+    {
+        guard let greeter = self.greeter as? Act else {
+            throw GreeterRemoteCallInterceptorError()
+        }
+
+        let anyReturn = try await withCheckedThrowingContinuation { (cc: CheckedContinuation<Any, Error>) in
+            let invocation = _invocation
+            Task {
+                var directDecoder = ClusterInvocationDecoder(invocation: invocation)
+                let directReturnHandler = ClusterInvocationResultHandler(directReturnContinuation: cc)
+
+                try await self.greeter.actorSystem.executeDistributedTarget(
+                    on: greeter,
+                    target: target,
+                    invocationDecoder: &directDecoder, handler: directReturnHandler
+                )
+            }
+        }
+
+        return anyReturn as! Res
+    }
+
+    func interceptRemoteCallVoid<Act, Err>(
+        on actor: Act,
+        target: RemoteCallTarget,
+        invocation _invocation: inout ClusterSystem.InvocationEncoder,
+        throwing: Err.Type
+    ) async throws
+        where Act: DistributedActor,
+        Act.ID == ActorID,
+        Err: Error
+    {
+        guard let greeter = self.greeter as? Act else {
+            throw GreeterRemoteCallInterceptorError()
+        }
+
+        _ = try await withCheckedThrowingContinuation { (cc: CheckedContinuation<Any, Error>) in
+            let invocation = _invocation
+            Task {
+                var directDecoder = ClusterInvocationDecoder(invocation: invocation)
+                let directReturnHandler = ClusterInvocationResultHandler(directReturnContinuation: cc)
+
+                try await self.greeter.actorSystem.executeDistributedTarget(
+                    on: greeter,
+                    target: target,
+                    invocationDecoder: &directDecoder, handler: directReturnHandler
+                )
+            }
+        }
+    }
+}
+
+private struct GreeterCodableError: Error, Codable {}
+private struct GreeterNonCodableError: Error {}
+private struct GreeterRemoteCallInterceptorError: Error, Codable {}
