+++ Tests/DistributedActorsTests/Metrics/ActorMemoryTests.swift
@@ -24,8 +24,8 @@ final class ActorMemoryTests: XCTestCase {
 
     func test_osx_actorShell_instanceSize() {
         #if os(OSX)
-        class_getInstanceSize(_ActorShell<Int>.self).shouldEqual(520)
-        class_getInstanceSize(_ActorShell<String>.self).shouldEqual(520)
+        class_getInstanceSize(_ActorShell<Int>.self).shouldEqual(608)
+        class_getInstanceSize(_ActorShell<String>.self).shouldEqual(608)
         #else
         print("Skipping test_osx_actorShell_instanceSize as requires Objective-C runtime")
         #endif
